+++ src/main/java/service/ConcurrentBank.java
@@ -2,9 +2,11 @@ package service;
 
 import entity.BankAccount;
 
+import java.math.BigDecimal;
+
 
 public interface ConcurrentBank {
-    Boolean transfer(BankAccount from, BankAccount to, int amount);
-    BankAccount createAccount(int balance);
-    int getTotalBalance();
+    Boolean transfer(BankAccount from, BankAccount to, BigDecimal amount);
+    BankAccount createAccount(BigDecimal balance);
+    BigDecimal getTotalBalance();
 }
